+++ tests/testthat/test_n_help.R
@@ -1,0 +1,4 @@
+test_that("config.help", {
+  x <- config.help()
+  expect_that(is.data.frame(x), equals(TRUE))
+})