+++ NEWS.md
@@ -1,3 +1,8 @@
+# configr v0.3.3
+
+- new function `str2config` to parse configuration string to R list object.
+- new function `config.help` to access external help resources, such as wiki, github, stackoverflow, etc.
+
 # configr v0.3.2.1
 
-- Support to write and convert json/ini/yaml to TOML format (Require python and python toml package)
+- Support to write and convert json/ini/yaml to TOML format (Require python and python toml package).